half base type
